backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Database (SQLCipher optional for MVP)
# sqlcipher3==0.5.2
# pysqlcipher3==1.2.0

# LLM Integration
ollama==0.1.7
requests==2.31.0
aiohttp==3.9.1

# Security & Encryption (flexible versions)
cryptography>=40.0.0
passlib==1.7.4
python-jose==3.3.0

# Configuration & Logging
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2

# Text Processing
# nltk==3.8.1  # Optional for MVP
tiktoken==0.5.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0

# Packaging & Distribution (for build only)
# pyinstaller==6.2.0
# setuptools==69.0.2
# wheel==0.42.0 